moved over the changes from wwrelay-utils

+++ debug_scripts/run_mbed_edge_core.sh
@@ -21,9 +21,27 @@ function run_edge_core() {
 	while true; do
         if ! pgrep -x "edge-core" > /dev/null
         then
-            $START_EDGE_CORE &
-            sleep 5
-            kill $(ps aux | grep 'mbed-devicejs-bridge' | awk '{print $2}');
+            # Only start edge-core if maestro is running
+            if pgrep "maestro" > /dev/null; then
+                #Give enough time for maestro to setup mcc_config and also acquire IP
+                sleep 30
+                $START_EDGE_CORE &
+                sleep 5
+                kill $(ps aux | grep 'mbed-devicejs-bridge' | awk '{print $2}');
+                kill $(ps aux | grep '/wigwag/mbed/pt-example' | awk '{print $2}');
+                sleep 5
+                /wigwag/mbed/pt-example -n pt-example --endpoint-postfix=-$(cat /sys/class/net/eth0/address) >> /var/log/pt-example.log 2>&1
+                # kill $(ps aux | grep '/wigwag/mbed/blept-example' | awk '{print $2}');
+                # sleep 5
+                # /etc/init.d/mept-ble start
+            fi
+        else
+            #edge-core is running, check if pt-example is, if not start it
+            if ! pgrep -x "pt-example" > /dev/null
+            then
+                sleep 30
+                /wigwag/mbed/pt-example -n pt-example --endpoint-postfix=-$(cat /sys/class/net/eth0/address) >> /var/log/pt-example.log 2>&1
+            fi
         fi
         sleep 5
 	done
